code/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.7
python-dotenv==1.0.0
selenium==4.8.2
